# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface LinkedInProfileInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  url: string;
  /** Must be true to authorize sending this URL to the profile-data provider. */
  consent: boolean;
}

export interface LinkedInProfile {
  /** @maxLength 100 */
  name: string;
  /**
     * @maxItems 1000
     * @items.minLength 1
     * @items.maxLength 50
     */
  skills: string[];
  /** @minimum 0 */
  detectedSkillCount: number;
}

export interface HealthStatus {
  status: string;
}

export type JobSearchInputEducation = typeof JobSearchInputEducation[keyof typeof JobSearchInputEducation];


export const JobSearchInputEducation = {
  'high-school': 'high-school',
  diploma: 'diploma',
  bachelors: 'bachelors',
  masters: 'masters',
  doctorate: 'doctorate',
} as const;

export type JobSearchInputWorkType = typeof JobSearchInputWorkType[keyof typeof JobSearchInputWorkType];


export const JobSearchInputWorkType = {
  remote: 'remote',
  hybrid: 'hybrid',
  onsite: 'onsite',
  any: 'any',
} as const;

export interface LocationPreference {
  /**
     * @minLength 2
     * @maxLength 100
     */
  country: string;
  /**
     * @maxItems 30
     * @items.minLength 2
     * @items.maxLength 100
     */
  cities: string[];
}

export type SkillProficiencyLevel = typeof SkillProficiencyLevel[keyof typeof SkillProficiencyLevel];


export const SkillProficiencyLevel = {
  beginner: 'beginner',
  intermediate: 'intermediate',
  advanced: 'advanced',
  expert: 'expert',
} as const;

export interface SkillProficiency {
  /**
     * @minLength 1
     * @maxLength 50
     */
  skill: string;
  level: SkillProficiencyLevel;
}

export interface JobSearchInput {
  /**
     * @minLength 2
     * @maxLength 100
     */
  title: string;
  /**
     * @minItems 1
     * @maxItems 10
     * @items.minLength 1
     * @items.maxLength 50
     */
  skills: string[];
  education: JobSearchInputEducation;
  /**
     * @minLength 2
     * @maxLength 100
     */
  location: string;
  /**
     * Alternative accepted locations. Empty cities means all cities in the country. Omit for legacy single-location searches.
     * @minItems 1
     * @maxItems 30
     */
  locations?: LocationPreference[];
  /**
     * User self-assessed proficiency. Does not change the existing skills match scoring.
     * @maxItems 10
     */
  skillLevels?: SkillProficiency[];
  workType: JobSearchInputWorkType;
}

export type JobMatchWorkType = typeof JobMatchWorkType[keyof typeof JobMatchWorkType];


export const JobMatchWorkType = {
  Remote: 'Remote',
  Hybrid: 'Hybrid',
  'On-site': 'On-site',
} as const;

export interface JobMatch {
  id: string;
  title: string;
  company: string;
  location: string;
  workType: JobMatchWorkType;
  source: string;
  url: string;
  publishedAt: string;
  /**
     * @minimum 0
     * @maximum 30
     */
  daysAgo: number;
  /**
     * @minimum 0
     * @maximum 100
     */
  score: number;
  matchedSkills: string[];
  summary: string;
  /** @nullable */
  salary?: string | null;
}

export interface JobSearchResponse {
  /** @maxItems 10 */
  jobs: JobMatch[];
  /** @minimum 0 */
  totalScanned: number;
  searchedAt: string;
  live: boolean;
}