# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: jobs
    description: Recent job matching
paths:
  /profiles/linkedin:
    post:
      operationId: importLinkedInProfile
      tags: [profiles]
      summary: Import available name and skills from a LinkedIn profile URL
      description: Sends the URL to People Data Labs. Requires explicit consent; successful matches may consume provider credits.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LinkedInProfileInput"
      responses:
        "200":
          description: Available profile fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LinkedInProfile"
        "400":
          description: Invalid URL or missing consent
        "404":
          description: No matching profile
        "422":
          description: No importable name or skills
        "429":
          description: Lookup limit reached
        "502":
          description: Provider unavailable or account requires attention
        "504":
          description: Provider lookup timed out
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /jobs/search:
    post:
      operationId: searchJobs
      tags: [jobs]
      summary: Find and rank recent jobs
      description: Searches recent public job listings and ranks the best matches.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobSearchInput"
      responses:
        "200":
          description: Ranked job matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobSearchResponse"
components:
  schemas:
    LinkedInProfileInput:
      type: object
      required: [url, consent]
      properties:
        url:
          type: string
          minLength: 1
          maxLength: 500
        consent:
          type: boolean
          description: Must be true to authorize sending this URL to the profile-data provider.
    LinkedInProfile:
      type: object
      required: [name, skills, detectedSkillCount]
      properties:
        name:
          type: string
          maxLength: 100
        skills:
          type: array
          maxItems: 1000
          items:
            type: string
            minLength: 1
            maxLength: 50
        detectedSkillCount:
          type: integer
          minimum: 0
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    JobSearchInput:
      type: object
      required: [title, skills, education, location, workType]
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 100
        skills:
          type: array
          minItems: 1
          maxItems: 10
          items:
            type: string
            minLength: 1
            maxLength: 50
        education:
          type: string
          enum: [high-school, diploma, bachelors, masters, doctorate]
        location:
          type: string
          minLength: 2
          maxLength: 100
        locations:
          description: Alternative accepted locations. Empty cities means all cities in the country. Omit for legacy single-location searches.
          type: array
          minItems: 1
          maxItems: 30
          items:
            $ref: "#/components/schemas/LocationPreference"
        skillLevels:
          description: User self-assessed proficiency. Does not change the existing skills match scoring.
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/SkillProficiency"
        workType:
          type: string
          enum: [remote, hybrid, onsite, any]
    LocationPreference:
      type: object
      required: [country, cities]
      properties:
        country:
          type: string
          minLength: 2
          maxLength: 100
        cities:
          type: array
          maxItems: 30
          uniqueItems: true
          items:
            type: string
            minLength: 2
            maxLength: 100
    SkillProficiency:
      type: object
      required: [skill, level]
      properties:
        skill:
          type: string
          minLength: 1
          maxLength: 50
        level:
          type: string
          enum: [beginner, intermediate, advanced, expert]
    JobMatch:
      type: object
      required: [id, title, company, location, workType, source, url, publishedAt, daysAgo, score, matchedSkills, summary]
      properties:
        id: { type: string }
        title: { type: string }
        company: { type: string }
        location: { type: string }
        workType:
          type: string
          enum: [Remote, Hybrid, On-site]
        source: { type: string }
        url: { type: string, format: uri }
        publishedAt: { type: string, format: date-time }
        daysAgo: { type: integer, minimum: 0, maximum: 30 }
        score: { type: integer, minimum: 0, maximum: 100 }
        matchedSkills:
          type: array
          items: { type: string }
        summary: { type: string }
        salary:
          type: [string, "null"]
    JobSearchResponse:
      type: object
      required: [jobs, totalScanned, searchedAt, live]
      properties:
        jobs:
          type: array
          maxItems: 10
          items:
            $ref: "#/components/schemas/JobMatch"
        totalScanned: { type: integer, minimum: 0 }
        searchedAt: { type: string, format: date-time }
        live: { type: boolean }