# lib/api-zod/src/generated/types/linkedInProfile.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LinkedInProfile {
  /** @maxLength 100 */
  name: string;
  /**
     * @maxItems 1000
     * @items.minLength 1
     * @items.maxLength 50
     */
  skills: string[];
  /** @minimum 0 */
  detectedSkillCount: number;
}