# lib/api-zod/src/generated/types/linkedInProfileInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface LinkedInProfileInput {
  /**
     * @minLength 1
     * @maxLength 500
     */
  url: string;
  /** Must be true to authorize sending this URL to the profile-data provider. */
  consent: boolean;
}